+++ src/index.ts
@@ -1,2 +1,5 @@
 export { BlueChipContract, contractAddresses } from './contracts';
 export { X0Sdk } from './x0-sdk';
+export { Erc721 } from './nft-manager';
+export { Erc721Alternative } from './nft-manager';
+export { type NftManager } from './nft-manager';
