+++ src/nft-manager/erc721-alternative.ts
@@ -24,6 +24,6 @@ export class Erc721Alternative implements NftManager {
 
   async tokenURI(tokenId: string): Promise<Record<string, any> | null> {
     const res = await this.alchemy.nft.getNftMetadata(this.contractAddress, tokenId);
-    return res;
+    return res.rawMetadata ?? null;
   }
 }
