+++ src/nft-manager/index.ts
@@ -1,0 +1,3 @@
+export { Erc721 } from './erc721';
+export { Erc721Alternative } from './erc721-alternative';
+export { type NftManager } from './interface';
