+++ src/x0-sdk.ts
@@ -1,8 +1,6 @@
 import { ethers, providers } from 'ethers';
 import { BlueChipContract, contractAddresses } from './contracts';
-import { Erc721 } from './nft-manager/erc721';
-import { Erc721Alternative } from './nft-manager/erc721-alternative';
-import { NftManager } from './nft-manager/interface';
+import { Erc721, Erc721Alternative, NftManager } from './nft-manager';
 
 export interface X0SdkConfig {
   /**
